# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PT Kledo Berhati Nyaman",
        "version": "0.1"
    },
    "paths": {
        "/api/settings": {
            "patch": {
                "tags": [
                    "Kledo Berhati Nyaman"
                ],
                "summary": "Edit Settings",
                "operationId": "0f706bcda4530ac1e7cff84cf43ab521",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "key": {
                                                "type": "string"
                                            },
                                            "value": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "key": "overtime_methods",
                                    "valye": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "key": {
                                            "type": "string",
                                            "example": "overrtime_methods"
                                        },
                                        "value": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/employee": {
            "post": {
                "tags": [
                    "Kledo Berhati Nyaman"
                ],
                "summary": "Store Employee",
                "operationId": "82ebe75d3e44c72fb1e12efb00b5a9df",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nama": {
                                                "type": "string"
                                            },
                                            "salary": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Saleh",
                                    "salary": "5000000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nama": {
                                            "type": "string",
                                            "example": "Saleh"
                                        },
                                        "salary": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/overtimes": {
            "post": {
                "tags": [
                    "Kledo Berhati Nyaman"
                ],
                "summary": "Store Overtimes",
                "operationId": "865b5449966dd4b24567802d6e9944b1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "employee_id": {
                                                "type": "integer"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "time_started": {
                                                "type": "time"
                                            },
                                            "time_ended": {
                                                "type": "time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "employee_id": "1",
                                    "date": "2022-02-02",
                                    "time_started": "10:10:00",
                                    "time_ended": "20:10:00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "employee_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "1"
                                        },
                                        "time_started": {
                                            "type": "time",
                                            "example": "10:10:00"
                                        },
                                        "time_ended": {
                                            "type": "time",
                                            "example": "20:10:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/overtime-pays/calculate": {
            "get": {
                "tags": [
                    "Kledo Berhati Nyaman"
                ],
                "summary": "Overtimes Pay Calculate",
                "operationId": "7eadf1347e73d2ba7fb26164f4f8f642",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "date": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "date": "2022-02"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "date": {
                                            "type": "string",
                                            "example": "2022-02"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}